Move constructor args to __call__() args
In order to effectively use a global ConfigOpts object, you
need to be able to initialize the global object with none of
the information we currently require at construction.

By moving those constructor args to the __call__() method,
we enable the global object usage model but also make the API
generally more flexible. For example, you can now reset the
object and re-use it for parsing a different set of config
files with the same options.

There are a couple of other minor behavior changes as a
result:

  - print_usage() and print_help() no longer works before
    the object has been called to parse options

  - registration of duplicate short options are no longer
    detected until the options are parsed

  - the --config-file and --config-dir options aren't
    registered until just before parsing the options since
    the default set of config files can be specified at
    that time

  - find_file() can't be used until after the options have
    been parsed, again because of the late registration
    of --config-file and --config-dir

Finally, an unregister_opt() method is added to support
the re-registeration of the --config-file and --config-dir
options.

Change-Id: I650d8e299e92cbc5d10da47a7ce1b73ca8066bd0