Split the dispatcher from the executor and server
Move the executors into a sub-package and organize the
code so that only one module imports eventlet.

Rename messaging/rpc/server.py to messaging/rpc/dispatcher.py and
leave only the dispatcher there.

Move the rest of the server code to messaging/server.py where it
can be reused with other dispatchers.

Remove the convenience functions for instantiating servers
to avoid having eventlet imported in the module with the base
class.

Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>