Make rpc.dispatcher private
This is just an implementation detail of the public EventletRPCServer
and BlockingRPCServer classes.

This is important because I'm not sure we've got the right separation
of concerns between executors and dispatchers yet:

  http://lists.openstack.org/pipermail/openstack-dev/2013-June/009934.html

  That implies that you need to pair a tulip-aware executor with a
  tulip-aware dispatcher. We'll probably need to do something similar for
  eventlet too.

I think we need the RPC dispatcher to just know about RPC specific stuff
and there's a single abstraction for I/O framework integration.

Change-Id: I5c43ec708b50d9be1abf6adaafe34f7917bf6d3a