Make RPCClient.check_for_lock a callable
We really don't want to depend on openstack.common.local since it
implies a dependency on eventlet.corolocal.

Instead, make the check_for_lock parameter a callable which is given the
ConfigOpts object and returns a list of held locks. To hook it up to
lockutils, just do:

  client = messaging.RPCClient(transport, target,
                               check_for_lock=lockutils.check_for_lock)

Although you probably want to use lockutils.debug_check_for_lock() which
only does the check if debugging is enabled.

Change-Id: Ic07ccdb92e3137c873f0ef791ba25625aa5803f1