Demo oslo.messaging based notify_decorator()
I don't think we'll add notify_decorator() to oslo.messaging, at least
initially. One reason for this is the way it probes the wrapped
function's arguments for a RequestContext. That's not behaviour we
should be encoding in oslo.messaging.

Another nice side-effect of keeping this in oslo-incubator is that Nova
encodes the path to this decorator in its monkey_patch_modules config
option and I we'll need to retain compat for existing configurations for
a deprecation period.

This patch shows how we might implement notify_decorator() with
oslo.messaging but we definitely need to revisit the fact that it
creates its own messaging transport rather than sharing the transport
that might be in use for RPC.

Change-Id: I04fbcb13f8d545c4a2a186602c32b08ca85fbd2f